fixed some issues

+++ lib/logic/update/anvil.dart
@@ -3,10 +3,10 @@ part of logic;
 void anvil() {
   grid.updateCell((cell, x, y) {
     final dir = cell.rot;
-    final double gravity = cell.data["gravity"];
-    final double velocity = cell.data["velocity"];
-    final double breakingVelocity = cell.data["breaking_velocity"];
-    final double lossUponLethalImpact = cell.data["impact_loss"];
-    final double terminalVelocity = cell.data["speed_limit"];
+    final double gravity = cell.data["gravity"].toDouble();
+    final double velocity = cell.data["velocity"].toDouble();
+    final double breakingVelocity = cell.data["breaking_velocity"].toDouble();
+    final double lossUponLethalImpact = cell.data["impact_loss"].toDouble();
+    final double terminalVelocity = cell.data["speed_limit"].toDouble();
   }, null, "anvil");
 }
